services/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
asyncpg==0.30.0
aiosqlite==0.20.0
redis==5.2.1
bcrypt==4.2.1
pyotp==2.9.0
cryptography==44.0.0
pydantic-settings==2.6.1
